+++ ts/features/bonus/cgn/hooks/useSpecialCtaCgn.tsx
@@ -7,7 +7,7 @@ import { cgnUnsubscribeSelector } from "../store/reducers/unsubscribe";
 import { fold, isLoading } from "../../../../common/model/RemoteValue";
 import I18n from "../../../../i18n";
 import * as analytics from "../../../services/common/analytics";
-import { ServiceId } from "../../../../../definitions/backend/ServiceId";
+import { ServiceId } from "../../../../../definitions/services/ServiceId";
 import { cgnUnsubscribe } from "../store/actions/unsubscribe";
 import { loadAvailableBonuses } from "../../common/store/actions/availableBonusesTypes";
 import { cgnActivationStart } from "../store/actions/activation";