+++ ts/store/reducers/backendStatus/remoteConfig.ts
@@ -175,13 +175,20 @@ export const isPremiumMessagesOptInOutEnabledSelector = createSelector(
  * if there is no data or the local Feature Flag is disabled,
  * false is the default value -> (CDC disabled)
  */
-export const isCdcEnabledSelector = createSelector(
+export const isCdcAppVersionSupportedSelector = createSelector(
   remoteConfigSelector,
   (remoteConfig): boolean =>
     cdcEnabled &&
     pipe(
       remoteConfig,
-      O.map(config => config.cdc.enabled),
+      O.map(config =>
+        isVersionSupported(
+          Platform.OS === "ios"
+            ? config.cdcV2.min_app_version.ios
+            : config.cdcV2.min_app_version.android,
+          getAppVersion()
+        )
+      ),
       O.getOrElse(() => false)
     )
 );
@@ -395,6 +402,58 @@ export const isPaymentsFeedbackBannerEnabledSelector = createSelector(
       O.getOrElse(() => false)
     )
 );
+
+export const isPaymentsWebViewFlowEnabledSelector = createSelector(
+  remoteConfigSelector,
+  (remoteConfig): boolean =>
+    pipe(
+      remoteConfig,
+      O.map(config =>
+        isVersionSupported(
+          Platform.OS === "ios"
+            ? config.newPaymentSection.webViewPaymentFlow?.min_app_version.ios
+            : config.newPaymentSection.webViewPaymentFlow?.min_app_version
+                .android,
+          getAppVersion()
+        )
+      ),
+      O.getOrElse(() => false)
+    )
+);
+
+/**
+ * Return the remote feature flag about the payment-method-specific psp banner enabled/disabled
+ * that is shown after a successful payment.
+ */
+export const isPaymentsPspBannerEnabledSelector = (paymentMethodName: string) =>
+  createSelector(remoteConfigSelector, (remoteConfig): boolean =>
+    pipe(
+      remoteConfig,
+      O.map(config =>
+        isVersionSupported(
+          Platform.OS === "ios"
+            ? config.newPaymentSection.pspBanner?.[paymentMethodName]
+                ?.min_app_version.ios
+            : config.newPaymentSection.pspBanner?.[paymentMethodName]
+                ?.min_app_version.android,
+          getAppVersion()
+        )
+      ),
+      O.getOrElse(() => false)
+    )
+  );
+
+/**
+ * Return the remote config about the payment-method-specific psp banner
+ */
+export const paymentsPspBannerConfigSelector = (paymentMethodName: string) =>
+  createSelector(remoteConfigSelector, (remoteConfig): Banner | undefined =>
+    pipe(
+      remoteConfig,
+      O.map(config => config.newPaymentSection.pspBanner?.[paymentMethodName]),
+      O.toUndefined
+    )
+  );
 
 /**
  * Return the remote config about the payment feedback banner